fixbug

+++ src/containers/accounts/components/bigAccount.tsx
@@ -6,6 +6,7 @@ import DropButtons from '@/components/dropButtons'
 import swal from 'sweetalert2'
 import { I18nCollectionAccount } from '@/i18n/i18n'
 import styles from '../accountsStyle'
+import { getShowName } from '@/utils'
 
 import Copy from '@/images/copy.png'
 // import Edit from '@/images/edit.png'
@@ -46,8 +47,18 @@ export class BigAccount extends React.Component<BigAccountProps> {
   removeAccount = () => {
     const {
       account: { id },
+      activeId,
       deleteAccount
     } = this.props
+    if (id === activeId) {
+      swal.fire({
+        showCloseButton: false,
+        icon: 'warning',
+        timer: 2000,
+        text: this.props.labels.deleteAccountWarning
+      })
+      return
+    }
     deleteAccount(id)
   }
 
@@ -58,13 +69,11 @@ export class BigAccount extends React.Component<BigAccountProps> {
       maximumFractionDigits: w
     })
   }
-  handleFunc = () => {
-    console.log('....')
-  }
 
   render() {
     const { classes, account, activeId, labels } = this.props
     const defaultName = account.name ? account.name : `${labels.account}${account.id}`
+    const showName = getShowName(defaultName)
     const btnArr = [
       {
         label: labels.modifyName,
@@ -81,7 +90,7 @@ export class BigAccount extends React.Component<BigAccountProps> {
           <DropButtons btnArray={btnArr} />
         </div>
         <div className={classes.bigAccountName}>
-          <span>{defaultName}</span>
+          <span>{showName}</span>
         </div>
         <div className={classNames(classes.bigId, { ['small-font']: String(account.id).length > 2 })}>{account.id}</div>
         <p className={classes.bigBalance} title={account.balance}>
